fix: move icons folder; feat: input atom - show/hide password
