+++ app/components/atoms/input/Input.tsx
@@ -1,6 +1,9 @@
 import styles from '../atoms.module.css';
+import eyeOpened from '@/public/icons/eye-solid.svg';
+import eyeClosed from '@/public/icons/eye-slash-solid.svg';
 
-import React, { InputHTMLAttributes } from 'react';
+import React, { InputHTMLAttributes, useState } from 'react';
+import Image from 'next/image';
 
 interface InputProps extends InputHTMLAttributes<HTMLInputElement> {
   changeFunc: (e: React.ChangeEvent<HTMLInputElement>) => void;
@@ -19,6 +22,14 @@ const Input = (inputProps: InputProps) => {
     value,
     ...rest
   } = inputProps;
+  const [showPassword, setShowPassword] = useState<boolean>(false);
+
+  const checkType = () => {
+    if (type === 'password') {
+      if (showPassword) return 'text';
+      else return 'password';
+    } else return type;
+  }
 
   return (
     <div className={styles[`${className}-section`]}>
@@ -33,12 +44,32 @@ const Input = (inputProps: InputProps) => {
       <input
         {...rest}
         value={value || ''}
-        type={type}
+        type={checkType()}
         id={id}
         onChange={changeFunc}
         accept={type === 'file' ? 'image/*' : ''}
         className={styles[`${className}`]}
       />
+      {type === 'password' && (
+        <span>
+          {value && showPassword === false && (
+            <Image
+            className={styles.icon}
+            src={eyeOpened}
+            alt={'eye-opened'}
+            onClick={() => setShowPassword(true)}
+          />
+          )}
+          {value && showPassword === true && (
+            <Image
+            className={styles.icon}
+            src={eyeClosed}
+            alt={'eye-opened'}
+            onClick={() => setShowPassword(false)}
+          />
+          )}
+        </span>
+      )}
     </div>
   );
 };
